python/requirements.txt
# Density Tracker Dependencies
websockets>=12.0
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0